Fix the problem with the pointers
Due to the parsing using std vectors the pointers are changing in every
push back. So the solution is to parse the nodes to the tree with only
the name and the level known (from tree file). And then from their
levels you can extract which node is parent of which.

So:

- Change the constructor of the Node to have only level and name as
  inputs
- Move the work for setting a parent to a separate function set_parent
- Fix the algorithm for find the parent of a node